Check that peer's host matches socket host
This is to prevent a peer from denying service to another peer by lying
about their end point address. The attacker would give the address
of the victim and then simply hold that heavyweight connection open, so
that requests by the victim would be rejected as crossed connections.
